MLECO-2135: Memory leak bug
The ApplicationContext::Set function allocates always new memory for the attibute.
When called multiple times (like it is done in most of the UseCaseHandler) this will generate a memory leak.

The function now checks if the attibute exists; If it does it frees the memory and then allocate memory for the new attribute.

Change-Id: I21db10009d6d0e360eab2dd33c344ef72eafe77f