Also download tiny_wav2letter
The wav2letter model being downloaded by Arm's script is too big.

This tiny model will fit into MRAM and work with the asr use-case, so
download it by default. It can be used with the asr use case by
changing the config setting asr_MODEL_TFLITE_PATH.

(It's not very good, and struggles with the decent test audio clips,
so I'm doubtful it will be worth wiring up to our live mics, but we
can try.)