MLECO-4935: Refactoring user options
Refactoring for CMake configuration options to allow target platform
to drive their own defaults for certain dependent options.

For example, the MPS3 target platform defaults to SSE-300 sub-system
while native target doesn't need to worry about sub-system and CMSIS
options which it will never use. This allows MPS4 target to default
to SSE-315 and set the NPU arch to Arm Ethos-U65 as it should, which
otherwise would have had to be explictly specified by the user.

A welcome by-product of this is also that the options summary at the
end of a successful CMake configuration stage would only include the
settings that are actually used by the given target.

Change-Id: Ibb7f8233d427cd33f39eeb4c71d88193aa998ebd
Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>
Tested-by: mlecosys <mlecosys@arm.com>
Reviewed-by: Alex Tawse <alex.tawse@arm.com>
Reviewed-by: Conor Kennedy <conor.kennedy@arm.com>
Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>