Increase minimum GCC HE stack+heap
Raise to minimum of 6K stack and 32K heap.

Heap choices were a mixture of 4K and 6K; 4K is cutting it fine when
adding printf trace in various places.

Newlib insists on 4K "page" alignment, and with the heap not being
aligned, the 32K of unaligned heap space could be not enough for some
our use cases, depending on where the heap start was.