MLECO-6389: Conditional generic logging for framework
TensorFlow Lite Micro based logging uses the overrides for
generic Arm Cortex-M CPU targets. This patch ensures that
this is the case only when the framework library target is
configured internally. This will allow external library
targets to be provided without mandating the use of generic
Arm Cortex-M target files being included in the build.

Change-Id: I022efc9d884d789e22f970e1f6e5ddb5e12efe4f
Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>