MLECO-6442: Display support for MPS4
Adding HDLCD driver for MPS4 platform and with wiring it
up to the newly refactored display component under HAL.

Splitting the CMakeLists under `display` component with
a CMakeLists file now present in MPS3 and MPS4 specific
subdirectories auto-included by the top level but their
targets "excluded from all" by default.

Resolution for MPS4 targets is kept the same as it is for
MPS3 boards to ensure the applications using the display
interface can continue working as they are. However, this
may be updated in future if we wish to use higher
resolution offered by the HDLCD component.

Change-Id: Ib1be3ff861267e8c560429969b5475e22899c4dc
Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>